backend/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
google-generativeai
pydantic
python-dotenv
